+++ packages/ui/src/screens/account_details/hooks.ts
@@ -31,6 +31,7 @@ const defaultTokenUnit: TokenUnit = {
 
 const initialState: AccountDetailState = {
   loading: true,
+  balanceLoading: true,
   exists: true,
   desmosProfile: null,
   overview: {
@@ -99,7 +100,7 @@ const formatBalance = (data: Data): BalanceType => {
   const rewardsAmount = formatToken(rewards, primaryTokenUnit);
 
   const commission = getDenom(
-    R.pathOr<NonNullable<NonNullable<(typeof data)['commission']>['coins']>>(
+    R.pathOr<NonNullable<NonNullable<typeof data['commission']>['coins']>>(
       [],
       ['commission', 'coins'],
       data
@@ -169,7 +170,7 @@ const formatOtherTokens = (data: Data) => {
     const availableRawAmount = getDenom(available, x);
     const availableAmount = formatToken(availableRawAmount.amount, x);
     const rewardsRawAmount = rewards.reduce((a, b) => {
-      const coins = R.pathOr<NonNullable<(typeof b)['coins']>>([], ['coins'], b);
+      const coins = R.pathOr<NonNullable<typeof b['coins']>>([], ['coins'], b);
       const denom = getDenom(coins, x);
       return Big(a).plus(denom.amount).toPrecision();
     }, '0');
@@ -196,7 +197,7 @@ const formatOtherTokens = (data: Data) => {
 // ==========================
 const formatAllBalance = (data: Data) => {
   const stateChange: Partial<AccountDetailState> = {
-    loading: false,
+    balanceLoading: false,
   };
 
   stateChange.rewards = formatRewards(data);
@@ -247,11 +248,11 @@ export const useAccountDetails = () => {
 
   useEffect(() => {
     const formattedRawData: {
-      commission?: (typeof commission)['commission'];
-      accountBalances?: (typeof available)['accountBalances'];
-      delegationBalance?: (typeof delegation)['delegationBalance'];
-      unbondingBalance?: (typeof unbonding)['unbondingBalance'];
-      delegationRewards?: (typeof rewards)['delegationRewards'];
+      commission?: typeof commission['commission'];
+      accountBalances?: typeof available['accountBalances'];
+      delegationBalance?: typeof delegation['delegationBalance'];
+      unbondingBalance?: typeof unbonding['unbondingBalance'];
+      delegationRewards?: typeof rewards['delegationRewards'];
     } = {};
     formattedRawData.commission = R.pathOr({ coins: [] }, ['commission'], commission);
     formattedRawData.accountBalances = R.pathOr({ coins: [] }, ['accountBalances'], available);