+++ packages/ui/src/screens/home/components/data_blocks/index.tsx
@@ -4,6 +4,7 @@ import { FC } from 'react';
 import SingleBlock from '@/screens/home/components/data_blocks/components/single_block';
 import { useDataBlocks } from '@/screens/home/components/data_blocks/hooks';
 import useStyles from '@/screens/home/components/data_blocks/styles';
+import CoinGeckoIcon from 'shared-utils/assets/icon-coingecko.svg';
 
 const DataBlocks: FC<ComponentDefault> = ({ className }) => {
   const { t } = useAppTranslation('home');
@@ -23,6 +24,8 @@ const DataBlocks: FC<ComponentDefault> = ({ className }) => {
     {
       key: t('price'),
       value: state.price !== null ? `$${numeral(state.price).format('0.000')}` : 'N/A',
+      description: t('dataFrom'),
+      Icon: <CoinGeckoIcon />,
       className: classes.price,
     },
     {
@@ -44,6 +47,7 @@ const DataBlocks: FC<ComponentDefault> = ({ className }) => {
           value={x.value}
           description={x.description}
           className={x.className}
+          Icon={x.Icon}
         />
       ))}
     </div>