+++ packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # shared-utils
 
+## 2.23.0
+
+### Minor Changes
+
+- [#1296](https://github.com/forbole/big-dipper-2.0-cosmos/pull/1296) [`584138bbb`](https://github.com/forbole/big-dipper-2.0-cosmos/commit/584138bbb2443b74d437546ad04327fa2f0dd003) Thanks [@rachelhox](https://github.com/rachelhox)! - feat: add Archway Big Dipper
+
+## 2.22.0
+
+### Minor Changes
+
+- [#1294](https://github.com/forbole/big-dipper-2.0-cosmos/pull/1294) [`99fb1bdf8`](https://github.com/forbole/big-dipper-2.0-cosmos/commit/99fb1bdf8be3e3baa022475903f0ebc8fc9f0619) Thanks [@rachelhox](https://github.com/rachelhox)! - add Jackal Big Dipper
+
+## 2.21.0
+
+### Minor Changes
+
+- [#1293](https://github.com/forbole/big-dipper-2.0-cosmos/pull/1293) [`d6d815915`](https://github.com/forbole/big-dipper-2.0-cosmos/commit/d6d815915a397c857247b32882222918eaef14e5) Thanks [@rachelhox](https://github.com/rachelhox)! - feat: add Coingecko attribution
+
 ## 2.20.0
 
 ### Minor Changes